Rename OpenMeteo client file
