+++ lib/open-meteo-client.ts
